tests/wpt/tests/html/canvas/offscreen/text/2d.text.draw.fontface.repeat.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.draw.fontface.repeat</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.draw.fontface.repeat</h1>
<p class="desc">Draw with the font immediately, then wait a bit until and draw again. (This crashes some version of WebKit.)</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var f = new FontFace("CanvasTest", "url('/fonts/CanvasTest.ttf')");
  f.load();
  document.fonts.add(f);
  await document.fonts.ready;
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.font = '67px CanvasTest';
  ctx.fillStyle = '#0f0';
  ctx.fillText('AA', 0, 50);

  await new Promise(resolve => t.step_timeout(resolve, 500));
  ctx.fillText('AA', 0, 50);
  _assertPixelApprox(canvas, 5,5, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 95,5, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 25,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 75,25, 0,255,0,255, 2);

}, "Draw with the font immediately, then wait a bit until and draw again. (This crashes some version of WebKit.)");
</script>
